Separate packet struct in another file
  Create `tcp_packet.h` for client and server.
