+++ src/teavpn2/include/teavpn2/client/linux/tcp.h
@@ -3,48 +3,10 @@
 #define __TEAVPN2__CLIENT__LINUX__TCP_H
 
 #include <stdint.h>
-#include <teavpn2/auth.h>
 #include <teavpn2/client/common.h>
-#include <teavpn2/client/linux/tcp.h>
+#include <teavpn2/server/linux/tcp_packet.h>
+#include <teavpn2/client/linux/tcp_packet.h>
 
-
-typedef enum __attribute__((packed)) {
-	CLI_PKT_HELLO	= 0,
-	CLI_PKT_AUTH	= 1,
-	CLI_PKT_DATA	= 2,
-	CLI_PKT_CLOSE	= 3,
-} cli_tcp_pkt_type;
-
-
-struct __attribute__((packed)) cli_tcp_pkt {
-	cli_tcp_pkt_type	type;
-	uint8_t			__pad;
-	uint16_t		length;
-	union {
-		char		raw_data[4096];
-		struct auth_pkt	auth;
-	};
-	uint8_t			__end;
-};
-
-STATIC_ASSERT(
-	sizeof(cli_tcp_pkt_type) == 1,
-	"Bad sizeof(cli_tcp_pkt_type)"
-);
-STATIC_ASSERT(
-	sizeof(struct cli_tcp_pkt) == (
-		  1	/* type      */
-		+ 1	/* __pad     */
-		+ 2	/* length    */
-		+ 4096	/* data      */
-		+ 1 	/* __end pad */
-	),
-	"Bad sizeof(struct cli_tcp_pkt)"
-);
-
-#define CLI_PKT_MIN_RSIZ (offsetof(struct cli_tcp_pkt, raw_data))
-#define CLI_PKT_END_OFF  (offsetof(struct cli_tcp_pkt, __end))
-#define CLI_PKT_DATA_SIZ (CLI_PKT_END_OFF - CLI_PKT_MIN_RSIZ)
 
 struct cli_tcp_state {
 	int			net_fd;
@@ -56,7 +18,8 @@ struct cli_tcp_state {
 	uint64_t		recv_c;
 	uint16_t		recv_s;
 	union {
-		char		recv_buf[1];
+		char			recv_buf[sizeof(struct srv_tcp_pkt)];
+		struct srv_tcp_pkt	srv_pkt;
 	};
 
 	uint64_t		send_c;
